Move serializer folder up #79
