+++ src/Chat/Infrastructure/Jms/Common.Domain.DomainEvent.yml
