+++ src/ConnectFour/Port/Adapter/Persistence/Jms/Common.Domain.DomainEvent.yml
