+++ src/ConnectFour/Port/Adapter/Persistence/Jms/ConnectFour.Domain.Game.Game.yml
