+++ src/ConnectFour/Port/Adapter/Persistence/Jms/ConnectFour.Domain.Game.State.State.yml
