+++ src/ConnectFour/Port/Adapter/Persistence/Jms/ConnectFour.Domain.Game.WinningRule.WinningRule.yml
