+++ src/ConnectFour/Port/Adapter/Persistence/Jms/FieldSubscriber.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Gaming\ConnectFour\Port\Adapter\Persistence\Jms\ConnectFour;
+namespace Gaming\ConnectFour\Port\Adapter\Persistence\Jms;
 
 use Gaming\ConnectFour\Domain\Game\Board\Field;
 use Gaming\ConnectFour\Domain\Game\Board\Point;
